e2e: move funs of framework/statefulset to e2e/apps & e2e/upgrades
