+++ test/e2e/framework/statefulset/wait.go
@@ -28,62 +28,6 @@ import (
 	"k8s.io/kubernetes/test/e2e/framework"
 )
 
-// WaitForPartitionedRollingUpdate waits for all Pods in set to exist and have the correct revision. set must have
-// a RollingUpdateStatefulSetStrategyType with a non-nil RollingUpdate and Partition. All Pods with ordinals less
-// than or equal to the Partition are expected to be at set's current revision. All other Pods are expected to be
-// at its update revision.
-func WaitForPartitionedRollingUpdate(c clientset.Interface, set *appsv1.StatefulSet) (*appsv1.StatefulSet, *v1.PodList) {
-	var pods *v1.PodList
-	if set.Spec.UpdateStrategy.Type != appsv1.RollingUpdateStatefulSetStrategyType {
-		framework.Failf("StatefulSet %s/%s attempt to wait for partitioned update with updateStrategy %s",
-			set.Namespace,
-			set.Name,
-			set.Spec.UpdateStrategy.Type)
-	}
-	if set.Spec.UpdateStrategy.RollingUpdate == nil || set.Spec.UpdateStrategy.RollingUpdate.Partition == nil {
-		framework.Failf("StatefulSet %s/%s attempt to wait for partitioned update with nil RollingUpdate or nil Partition",
-			set.Namespace,
-			set.Name)
-	}
-	WaitForState(c, set, func(set2 *appsv1.StatefulSet, pods2 *v1.PodList) (bool, error) {
-		set = set2
-		pods = pods2
-		partition := int(*set.Spec.UpdateStrategy.RollingUpdate.Partition)
-		if len(pods.Items) < int(*set.Spec.Replicas) {
-			return false, nil
-		}
-		if partition <= 0 && set.Status.UpdateRevision != set.Status.CurrentRevision {
-			framework.Logf("Waiting for StatefulSet %s/%s to complete update",
-				set.Namespace,
-				set.Name,
-			)
-			SortStatefulPods(pods)
-			for i := range pods.Items {
-				if pods.Items[i].Labels[appsv1.StatefulSetRevisionLabel] != set.Status.UpdateRevision {
-					framework.Logf("Waiting for Pod %s/%s to have revision %s update revision %s",
-						pods.Items[i].Namespace,
-						pods.Items[i].Name,
-						set.Status.UpdateRevision,
-						pods.Items[i].Labels[appsv1.StatefulSetRevisionLabel])
-				}
-			}
-			return false, nil
-		}
-		for i := int(*set.Spec.Replicas) - 1; i >= partition; i-- {
-			if pods.Items[i].Labels[appsv1.StatefulSetRevisionLabel] != set.Status.UpdateRevision {
-				framework.Logf("Waiting for Pod %s/%s to have revision %s update revision %s",
-					pods.Items[i].Namespace,
-					pods.Items[i].Name,
-					set.Status.UpdateRevision,
-					pods.Items[i].Labels[appsv1.StatefulSetRevisionLabel])
-				return false, nil
-			}
-		}
-		return true, nil
-	})
-	return set, pods
-}
-
 // WaitForRunning waits for numPodsRunning in ss to be Running and for the first
 // numPodsReady ordinals to be Ready.
 func WaitForRunning(c clientset.Interface, numPodsRunning, numPodsReady int32, ss *appsv1.StatefulSet) {
@@ -130,19 +74,6 @@ func WaitForState(c clientset.Interface, ss *appsv1.StatefulSet, until func(*app
 	}
 }
 
-// WaitForStatus waits for the StatefulSetStatus's ObservedGeneration to be greater than or equal to set's Generation.
-// The returned StatefulSet contains such a StatefulSetStatus
-func WaitForStatus(c clientset.Interface, set *appsv1.StatefulSet) *appsv1.StatefulSet {
-	WaitForState(c, set, func(set2 *appsv1.StatefulSet, pods *v1.PodList) (bool, error) {
-		if set2.Status.ObservedGeneration >= set.Generation {
-			set = set2
-			return true, nil
-		}
-		return false, nil
-	})
-	return set
-}
-
 // WaitForRunningAndReady waits for numStatefulPods in ss to be Running and Ready.
 func WaitForRunningAndReady(c clientset.Interface, numStatefulPods int32, ss *appsv1.StatefulSet) {
 	WaitForRunning(c, numStatefulPods, numStatefulPods, ss)
@@ -162,65 +93,6 @@ func WaitForPodReady(c clientset.Interface, set *appsv1.StatefulSet, podName str
 		return false, nil
 	})
 	return set, pods
-}
-
-// WaitForPodNotReady waits for the Pod named podName in set to exist and to not have a Ready condition.
-func WaitForPodNotReady(c clientset.Interface, set *appsv1.StatefulSet, podName string) (*appsv1.StatefulSet, *v1.PodList) {
-	var pods *v1.PodList
-	WaitForState(c, set, func(set2 *appsv1.StatefulSet, pods2 *v1.PodList) (bool, error) {
-		set = set2
-		pods = pods2
-		for i := range pods.Items {
-			if pods.Items[i].Name == podName {
-				return !podutil.IsPodReady(&pods.Items[i]), nil
-			}
-		}
-		return false, nil
-	})
-	return set, pods
-}
-
-// WaitForRollingUpdate waits for all Pods in set to exist and have the correct revision and for the RollingUpdate to
-// complete. set must have a RollingUpdateStatefulSetStrategyType.
-func WaitForRollingUpdate(c clientset.Interface, set *appsv1.StatefulSet) (*appsv1.StatefulSet, *v1.PodList) {
-	var pods *v1.PodList
-	if set.Spec.UpdateStrategy.Type != appsv1.RollingUpdateStatefulSetStrategyType {
-		framework.Failf("StatefulSet %s/%s attempt to wait for rolling update with updateStrategy %s",
-			set.Namespace,
-			set.Name,
-			set.Spec.UpdateStrategy.Type)
-	}
-	WaitForState(c, set, func(set2 *appsv1.StatefulSet, pods2 *v1.PodList) (bool, error) {
-		set = set2
-		pods = pods2
-		if len(pods.Items) < int(*set.Spec.Replicas) {
-			return false, nil
-		}
-		if set.Status.UpdateRevision != set.Status.CurrentRevision {
-			framework.Logf("Waiting for StatefulSet %s/%s to complete update",
-				set.Namespace,
-				set.Name,
-			)
-			SortStatefulPods(pods)
-			for i := range pods.Items {
-				if pods.Items[i].Labels[appsv1.StatefulSetRevisionLabel] != set.Status.UpdateRevision {
-					framework.Logf("Waiting for Pod %s/%s to have revision %s update revision %s",
-						pods.Items[i].Namespace,
-						pods.Items[i].Name,
-						set.Status.UpdateRevision,
-						pods.Items[i].Labels[appsv1.StatefulSetRevisionLabel])
-				}
-			}
-			return false, nil
-		}
-		return true, nil
-	})
-	return set, pods
-}
-
-// WaitForRunningAndNotReady waits for numStatefulPods in ss to be Running and not Ready.
-func WaitForRunningAndNotReady(c clientset.Interface, numStatefulPods int32, ss *appsv1.StatefulSet) {
-	WaitForRunning(c, numStatefulPods, 0, ss)
 }
 
 // WaitForStatusReadyReplicas waits for the ss.Status.ReadyReplicas to be equal to expectedReplicas
